Default and Red Eyes
The main Eyeballs script contains code that looks at the object it's a component of and checks whether there is a component under the name DefaultEyes or RedEyes depending on whether 1 or 2 is pressed on the keyboard. It then changes the sorting layer of the object respectively based on whether it should be hidden or seen for that eyeball. For this to work, all the objects I tested were set to the background sorting layer and then switched between sorting layer 1 and 0 to determine whether they were seen or hidden.