Eyeball Script Efficiency
Changed the implementation of Eyeballs.cs so that each sprite has a list of bools that can be set to true/false based on whether the object should be seen for certain eyeballs. Right now, default eyes is Element 0 (pressing 1 on the keyboard) and red eyes is Element 1 (pressing 2 on the keyboard). This removes the need for having additional scripts that the Eyeballs.cs script will look for as components of the sprites. Also changed the implementation to change the color of the sprite from transparent to nontransparent rather than changing its order in the sorting layer.